Functioning express app.get for both directory list and file
