+++ lib/app.js
@@ -1,20 +1,28 @@
 const express = require('express');
+const path = require('path');
 const app = express();
+
 const handlers = require('./requestHandlers');
+const filePath = path.join(__dirname, '/store');
 
-module.exports = function(request, response) {
-  console.log('app in effect...');
-  var method = request.method;
+module.exports = app;
 
-  console.log('Request for ' + request.pathname + ' received');
-  console.log(method);
+console.log('app in effect on path', filePath);
 
-  app.get('/lib/store', (request, response) => {
-    console.log('calling handlers...');
-    handlers.get(request, response);
-  });
+app.get('/favicon.ico', (request, response) => { 
+  const favicon = fs.createReadStream('favicon.ico');
+  favicon.pipe(response);
+});
 
-  app.use(errorHandler);
+app.get('/store', (request, response) => {
+  console.log('You requested ', request.path);
+  handlers.getDir(filePath, request, response);
+});
+
+app.get('/store/:fileName', (request, response) => {
+  console.log('You requested ', request.path);
+  handlers.getFile(filePath + '/' + request.params.fileName, request, response);
+});
   
   // switch(method) {
   // case 'GET': 
@@ -32,9 +40,12 @@ module.exports = function(request, response) {
   // default:
   //   handlers.fourohfour(request, response);
   // }
-};
 
-function errorHandler(error, request, response, next) {
-  console.log('In error handler, ', error);
-  next(error);
-};
+
+
+app.use('*', (request, response, next) => {
+  console.log('UNAUTHORIZED');
+  response.write('404 Not Found. Please visit localhost:3000/lib/store for a list of available documents.');
+  response.end();
+  next();
+});
