+++ lib/bodyParser.js
@@ -1,0 +1,15 @@
+module.exports = function() {
+  var body = '';
+
+  request.on('data', (chunk) => {
+    body += chunk;
+  });
+
+  request.on('end', () => {
+    if(body) {
+      request.body = JSON.parse(body);
+      console.log('body: ' + body);
+    }
+  });
+};
+
