+++ lib/requestHandlers.js
@@ -3,40 +3,28 @@ const path = require('path');
 const fs = require('fs');
 const handlers = {};
 
-handlers.get = function(request, response) {
-  console.log('handlers handling');
-  const pathArray = request.pathname.split('/');
-  const dirThing = (filePath) => {
-    sander.readdir(filePath)
-      .then((data) => {
-        response.write('You requested ' + filePath.substr(1) + '\n');
-        data.forEach((el) => {
-          response.write(el + '\n');
-        });
-        response.end();
-      })
-      .catch((err) => {
-        console.log(err);
+handlers.getDir = function(filePath, request, response) {
+  sander.readdir(filePath)
+    .then((data) => {
+      response.write('You requested ' + filePath + '\n');
+      data.forEach((el) => {
+        response.write(el + '\n');
       });
-  };
+      response.end();
+    })
+    .catch((err) => {
+      console.log(err);
+    });
+};
 
-  const fileThing = (filePath) => {
-    console.log('filePath: ' + filePath);
-    console.log('pathArray: ', pathArray);
-    sander.readFile(filePath)
-      .then((data) => {
-        response.write('You requested ' + filePath.substr(1) + '\n');
-        response.write(data);
-        response.end();
-      });
-  };
-
-  if (pathArray[pathArray.length - 1] && pathArray[pathArray.length - 1] !== 'store') {
-    fileThing('.' + request.pathname);
-    return;
-  }
-  dirThing('.' + request.pathname);
-
+handlers.getFile = function(fileName, request, response) {
+  console.log('fileName: ' + fileName);
+  sander.readFile(fileName)
+    .then((data) => {
+      response.write('You requested ' + fileName + '\n');
+      response.write(data);
+      response.end();
+    });
 };
 
 handlers.post = function(request, response) {
